// src/app/interfaces/decoderData.interface.ts
export interface IDecoderData {
  uuid?: string,
  commonName?: string;
  issuerName?: string;
  startDate?: string;
  endDate?: string;
  organizationName?: string;
  organizationalUnitName?: string;
  serialNumber?: string;
  localityName?: string;
  stateOrProvinceName?: string;
  title?: string;
  surname?: string;
  givenName?: string;
  active?: boolean;
}